requirements.txt
requests
python-dotenv
telethon==1.31.1